[4254] Make explorer services reusable
Bug: https://github.com/eclipse-sirius/sirius-web/issues/4254
