+++ packages/sirius-web/backend/sirius-web-application/src/main/java/org/eclipse/sirius/web/application/studio/services/representations/DomainExplorerServices.java
@@ -17,28 +17,19 @@ import java.util.Collection;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 
-import org.eclipse.emf.common.util.URI;
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.EStructuralFeature.Setting;
 import org.eclipse.emf.ecore.InternalEObject;
 import org.eclipse.emf.ecore.resource.Resource;
-import org.eclipse.emf.ecore.resource.ResourceSet;
-import org.eclipse.emf.edit.domain.EditingDomain;
 import org.eclipse.sirius.components.collaborative.api.IRepresentationImageProvider;
-import org.eclipse.sirius.components.core.CoreImageConstants;
-import org.eclipse.sirius.components.core.RepresentationMetadata;
 import org.eclipse.sirius.components.core.api.IEditingContext;
 import org.eclipse.sirius.components.core.api.IObjectService;
 import org.eclipse.sirius.components.core.api.IURLParser;
-import org.eclipse.sirius.components.core.api.SemanticKindConstants;
 import org.eclipse.sirius.components.domain.Domain;
 import org.eclipse.sirius.components.domain.Entity;
-import org.eclipse.sirius.components.emf.ResourceMetadataAdapter;
-import org.eclipse.sirius.components.emf.services.JSONResourceFactory;
-import org.eclipse.sirius.components.emf.services.api.IEMFEditingContext;
 import org.eclipse.sirius.web.application.UUIDParser;
+import org.eclipse.sirius.web.application.views.explorer.services.api.IExplorerServices;
 import org.eclipse.sirius.web.domain.boundedcontexts.representationdata.services.api.IRepresentationMetadataSearchService;
 import org.springframework.data.jdbc.core.mapping.AggregateReference;
 import org.springframework.stereotype.Service;
@@ -61,54 +52,35 @@ public class DomainExplorerServices {
 
     private final IRepresentationMetadataSearchService representationMetadataSearchService;
 
-    private final IURLParser urlParser;
+    private final IExplorerServices explorerServices;
 
-    private final List<IRepresentationImageProvider> representationImageProviders;
-
-    public DomainExplorerServices(IObjectService objectService, IRepresentationMetadataSearchService representationMetadataSearchService, IURLParser urlParser, List<IRepresentationImageProvider> representationImageProviders) {
+    public DomainExplorerServices(IObjectService objectService, IRepresentationMetadataSearchService representationMetadataSearchService, IURLParser urlParser, List<IRepresentationImageProvider> representationImageProviders, IExplorerServices explorerServices) {
         this.objectService = Objects.requireNonNull(objectService);
         this.representationMetadataSearchService = Objects.requireNonNull(representationMetadataSearchService);
-        this.urlParser = Objects.requireNonNull(urlParser);
-        this.representationImageProviders = Objects.requireNonNull(representationImageProviders);
+        this.explorerServices = Objects.requireNonNull(explorerServices);
     }
 
     public String getKind(Object self) {
         String kind = "";
-        if (self instanceof RepresentationMetadata representationMetadata) {
-            kind = representationMetadata.kind();
-        } else if (self instanceof Resource) {
-            kind = DOCUMENT_KIND;
-        } else if (self instanceof SettingWrapper) {
+        if (self instanceof SettingWrapper) {
             kind = "setting";
         } else {
-            kind = this.objectService.getKind(self);
+            kind = this.explorerServices.getKind(self);
         }
         return kind;
     }
 
     public boolean hasChildren(IEditingContext editingContext, Object self) {
         boolean hasChildren = false;
-        if (self instanceof Resource resource) {
-            hasChildren = !resource.getContents().isEmpty();
-        } else if (self instanceof EObject eObject) {
-            hasChildren = !eObject.eContents().isEmpty();
-
-            if (!hasChildren) {
-                var optionalEditingContextId = new UUIDParser().parse(editingContext.getId());
-                if (optionalEditingContextId.isPresent()) {
-                    var projectId = optionalEditingContextId.get();
-                    String id = this.objectService.getId(eObject);
-                    hasChildren = this.representationMetadataSearchService.existAnyRepresentationForProjectAndTargetObjectId(AggregateReference.to(projectId), id);
-                }
-            }
-
-            if (!hasChildren && self instanceof Entity) {
-                hasChildren = true;
-            }
-        } else if (self instanceof SettingWrapper wrapper) {
+        if (self instanceof SettingWrapper wrapper) {
             var value = wrapper.setting.get(true);
             if (value instanceof Collection<?> collection) {
                 hasChildren = !collection.isEmpty();
+            }
+        } else {
+            hasChildren = this.explorerServices.hasChildren(self, editingContext);
+            if (!hasChildren && self instanceof Entity) {
+                hasChildren = true;
             }
         }
         return hasChildren;
@@ -116,19 +88,9 @@ public class DomainExplorerServices {
 
     public List<Object> getElements(IEditingContext editingContext) {
         List<Object> elements = new ArrayList<>();
-
-        var optionalResourceSet = Optional.of(editingContext).filter(IEditingContext.class::isInstance)
-                .filter(IEMFEditingContext.class::isInstance)
-                .map(IEMFEditingContext.class::cast)
-                .map(IEMFEditingContext::getDomain)
-                .map(EditingDomain::getResourceSet);
-
-        var resources = optionalResourceSet.map(resourceSet -> resourceSet.getResources().stream()
+        this.explorerServices.getDefaultElements(editingContext).stream()
                 .filter(resource -> resource.getContents().stream().anyMatch(Domain.class::isInstance))
-                .sorted(Comparator.nullsLast(Comparator.comparing(this::getResourceLabel, String.CASE_INSENSITIVE_ORDER)))
-                .toList()
-        ).orElseGet(ArrayList::new);
-        elements.addAll(resources);
+                .forEach(elements::add);
         return elements;
     }
 
@@ -166,53 +128,31 @@ public class DomainExplorerServices {
 
     public Object getParent(Object self, IEditingContext editingContext, String treeItemId) {
         Object result = null;
-
-        if (self instanceof RepresentationMetadata) {
-            var optionalRepresentationMetadata = new UUIDParser().parse(treeItemId).flatMap(this.representationMetadataSearchService::findMetadataById);
-            var repId = optionalRepresentationMetadata.map(org.eclipse.sirius.web.domain.boundedcontexts.representationdata.RepresentationMetadata::getTargetObjectId).orElse(null);
-            result = this.objectService.getObject(editingContext, repId);
-        } else if (self instanceof EObject eObject) {
-            Object semanticContainer = eObject.eContainer();
-            if (semanticContainer == null) {
-                semanticContainer = eObject.eResource();
-            }
-            result = semanticContainer;
-        } else if (self instanceof SettingWrapper wrapper) {
-            // the parent of the superTypes node is the object associated to this Setting
+        if (self instanceof SettingWrapper wrapper) {
+         // the parent of the superTypes node is the object associated to this Setting
             result = wrapper.setting.getEObject();
+        } else {
+            result = this.explorerServices.getParent(self, treeItemId, editingContext);
         }
         return result;
     }
 
     public String getTreeItemLabel(Object self) {
         String label = "";
-
-        if (self instanceof RepresentationMetadata representationMetadata) {
-            label = representationMetadata.label();
-        } else if (self instanceof Resource resource) {
-            label = this.getResourceLabel(resource);
-        } else if (self instanceof EObject) {
-            label = this.objectService.getLabel(self);
-            if (label.isBlank()) {
-                var kind = this.objectService.getKind(self);
-                label = this.urlParser.getParameterValues(kind).get(SemanticKindConstants.ENTITY_ARGUMENT).get(0);
-            }
-        } else if (self instanceof SettingWrapper wrapper) {
+        if (self instanceof SettingWrapper wrapper) {
             label = wrapper.setting.getEStructuralFeature().getName();
+        } else {
+            label = this.explorerServices.getLabel(self);
         }
         return label;
     }
 
     public String getTreeItemId(Object self) {
         String id = null;
-        if (self instanceof RepresentationMetadata representationMetadata) {
-            id = representationMetadata.id();
-        } else if (self instanceof Resource resource) {
-            id = resource.getURI().path().substring(1);
-        } else if (self instanceof EObject) {
-            id = this.objectService.getId(self);
-        } else if (self instanceof SettingWrapper wrapper) {
+        if (self instanceof SettingWrapper wrapper) {
             id = SETTING + this.objectService.getId(wrapper.setting.getEObject()) + SETTING_ID_SEPARATOR + wrapper.setting.getEStructuralFeature().getName();
+        } else {
+            id = this.explorerServices.getTreeItemId(self);
         }
         return id;
     }
@@ -230,47 +170,14 @@ public class DomainExplorerServices {
                 result = new SettingWrapper(internalObject.eSetting(internalObject.eClass().getEStructuralFeature(featureName)));
             }
         } else {
-            var optionalObject = this.objectService.getObject(editingContext, treeItemId);
-            if (optionalObject.isPresent()) {
-                result = optionalObject.get();
-            } else {
-                var optionalEditingDomain = Optional.of(editingContext)
-                        .filter(IEMFEditingContext.class::isInstance)
-                        .map(IEMFEditingContext.class::cast)
-                        .map(IEMFEditingContext::getDomain);
-
-                if (optionalEditingDomain.isPresent()) {
-                    var editingDomain = optionalEditingDomain.get();
-                    ResourceSet resourceSet = editingDomain.getResourceSet();
-                    URI uri = new JSONResourceFactory().createResourceURI(treeItemId);
-
-                    result = resourceSet.getResources().stream()
-                            .filter(resource -> resource.getURI().equals(uri))
-                            .findFirst()
-                            .orElse(null);
-                }
-            }
+            result = this.explorerServices.getTreeItemObject(treeItemId, editingContext);
         }
 
         return result;
     }
 
     public List<String> getIconURL(Object self) {
-        List<String> result = List.of(CoreImageConstants.DEFAULT_SVG);
-        if (self instanceof EObject) {
-            result = this.objectService.getImagePath(self);
-        } else if (self instanceof RepresentationMetadata representationMetadata) {
-            if (!representationMetadata.iconURLs().isEmpty()) {
-                result = representationMetadata.iconURLs();
-            } else {
-                result = this.representationImageProviders.stream()
-                        .flatMap(provider -> provider.getImageURL(representationMetadata.kind()).stream())
-                        .toList();
-            }
-        } else if (self instanceof Resource) {
-            result = List.of("/explorer/Resource.svg");
-        }
-        return result;
+        return this.explorerServices.getImageURL(self);
     }
 
     public boolean isDeletable(Object self) {
@@ -283,15 +190,6 @@ public class DomainExplorerServices {
 
     public boolean isSelectable(Object self) {
         return !(self instanceof SettingWrapper);
-    }
-
-    private String getResourceLabel(Resource resource) {
-        return resource.eAdapters().stream()
-                .filter(ResourceMetadataAdapter.class::isInstance)
-                .map(ResourceMetadataAdapter.class::cast)
-                .findFirst()
-                .map(ResourceMetadataAdapter::getName)
-                .orElse(resource.getURI().lastSegment());
     }
 
     public Object toggleAbstractEntity(Object self) {
